refactor: move timestamp tests over from process-event.ts
